pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "treehacks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.13"

dependencies = [
  "eclipse-zenoh>=1.7.2",
  "matplotlib>=3.10.8",
  "mujoco[all]>=3.5.0",
  "numpy>=2.4.2",
  "onnxruntime>=1.24.1",
  "scikit-learn>=1.8.0",
  "scipy>=1.17.0",
  "torch>=2.9.1,<2.10.0",
  "torchaudio>=2.9.1,<2.10.0",
  "torchvision>=0.24.0,<0.25.0",
  "triton>=3.5.0,<3.6.0",
  "unitree-sdk2py",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.sources]
unitree-sdk2py = { git = "https://github.com/Kaweees/unitree_sdk2_python.git", branch = "main" }
torch = [
  { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
  { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
triton = [
  { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[dependency-groups]
dev = ["pre-commit", "ipykernel"]

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
